SUPPORT REGISTRATION WITH CREDIT CARD INFO
User module needed to be extended and support for credit card details
must be implemented during user creation. This would allow users
to be able to use a default payment method

Edited user.js module to take optional parameters and write to database
if they are not empty. Updated index.js to pass needed parameters from
form to user.register() and if they are not empty write them to a database.