ADDED TICKET MODULE
Created a ticket module to allow us to easily add tickets to the db,
along with information about the movie name, buyer and price. This will
hopefully be extended to add more functionality for the more advanced
tasks

Has a similar structure to user.js, changes were made so that it becomes
more suitable.