ADDED A CART ICON TO THE NAVBAR ONLY APPEARS IF THE USER IS LOGGED IN.
Added a cart icon where everything the user selects will be sent to that
page so he can see what he is about to pay for.
As well as the CSS for that.

This will make the Navigation bar a bit more interesting to look at and
saves the user from having to read since the shopping cart icon
is a common Icon thats used to store things.