ADDED CART MODULE
Added new shoppingCart module that will allow for the creation of a cart
using the map datatype. Called constructor within /login 's post method

Will allow for us to easily add and remove items from the cart, and
will hopefully make it easier to write all the purchased tickets to the
db once it's extended.