ADDED PRICE BANDS TO THE TICKETS
The ticket page now displays the tickets remaining in each price band
(low, medium or high). The user is able to choose up  to the maximum
possible from each category. The user is then redirected to the payment
page that adds up the tickets submitted from the form on the ticket page

Added a new method to the ticket module that returns the 'COUNT(Id)' of
the specified movie, date, time and price band. There are 3 SQL
statements, each returning for the said price band.
There is a 'reduce()' function that then adds up the tickets and posts
them on the payment page.